requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
python-dotenv==1.0.1
gunicorn==22.0.0
